.Net: Fixing Processes food samples diagram links + step name fix (#9262)
### Description

- Fixing Processes Food Samples links to respective diagrams + adding
some steps specific names
- Reorganizing Process Food Samples in 2 files -> allows to build
concepts in more detail later: more samples to be added in another PR in
the near future

### Contribution Checklist

<!-- Before submitting this PR, please make sure: -->

- [x] The code builds clean without any errors or warnings
- [x] The PR follows the [SK Contribution
Guidelines](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md)
and the [pre-submission formatting
script](https://github.com/microsoft/semantic-kernel/blob/main/CONTRIBUTING.md#development-scripts)
raises no violations
- [x] All unit tests pass, and I have added new tests where possible
- [x] I didn't break anyone :smile:
