+++ dotnet/samples/GettingStartedWithProcesses/Step03/Processes/PotatoFriesProcess.cs
@@ -8,7 +8,7 @@ namespace Step03.Processes;
 
 /// <summary>
 /// Sample process that showcases how to create a process with sequential steps and reuse of existing steps.<br/>
-/// For a visual reference of the FriedFishProcess check this <see href="https://github.com/microsoft/semantic-kernel/tree/main/dotnet/samples/GettingStartedWithProcesses/README.md#Potato_Fries_Preparation_Process" >diagram</see>
+/// For a visual reference of the FriedFishProcess check this <see href="https://github.com/microsoft/semantic-kernel/blob/main/dotnet/samples/GettingStartedWithProcesses/README.md#potato-fries-preparation-process" >diagram</see>
 /// </summary>
 public static class PotatoFriesProcess
 {
@@ -26,7 +26,7 @@ public static class PotatoFriesProcess
         var processBuilder = new ProcessBuilder(processName);
 
         var gatherIngredientsStep = processBuilder.AddStepFromType<GatherPotatoFriesIngredientsStep>();
-        var sliceStep = processBuilder.AddStepFromType<CutFoodStep>();
+        var sliceStep = processBuilder.AddStepFromType<CutFoodStep>("sliceStep");
         var fryStep = processBuilder.AddStepFromType<FryFoodStep>();
 
         processBuilder
