+++ dotnet/samples/GettingStartedWithProcesses/Step03/Step03a_FoodPreparation.cs
@@ -1,17 +1,16 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 
 using Microsoft.SemanticKernel;
-using Step03.Models;
 using Step03.Processes;
 
 namespace Step03;
 
 /// <summary>
 /// Demonstrate creation of <see cref="KernelProcess"/> and
-/// eliciting its response to three explicit user messages.
-/// For visual reference of the processes used here check the diagram in: https://github.com/microsoft/semantic-kernel/tree/main/dotnet/samples/GettingStartedWithProcesses/README.md#step03_foodPreparation
+/// eliciting different food related events.
+/// For visual reference of the processes used here check the diagram in: https://github.com/microsoft/semantic-kernel/tree/main/dotnet/samples/GettingStartedWithProcesses/README.md#step03a_foodPreparation
 /// </summary>
-public class Step03_FoodPreparation(ITestOutputHelper output) : BaseTest(output, redirectSystemConsoleOutput: true)
+public class Step03a_FoodPreparation(ITestOutputHelper output) : BaseTest(output, redirectSystemConsoleOutput: true)
 {
     // Target Open AI Services
     protected override bool ForceOpenAI => true;
@@ -56,42 +55,6 @@ public class Step03_FoodPreparation(ITestOutputHelper output) : BaseTest(output,
         using var runningProcess = await kernelProcess.StartAsync(kernel, new KernelProcessEvent()
         {
             Id = externalTriggerEvent, Data = new List<string>()
-        });
-    }
-
-    [Fact]
-    public async Task UseSingleOrderFriedFishAsync()
-    {
-        await UsePrepareFoodOrderProcessSingleItemAsync(FoodItem.FriedFish);
-    }
-
-    [Fact]
-    public async Task UseSingleOrderPotatoFriesAsync()
-    {
-        await UsePrepareFoodOrderProcessSingleItemAsync(FoodItem.PotatoFries);
-    }
-
-    [Fact]
-    public async Task UseSingleOrderFishSandwichAsync()
-    {
-        await UsePrepareFoodOrderProcessSingleItemAsync(FoodItem.FishSandwich);
-    }
-
-    [Fact]
-    public async Task UseSingleOrderFishAndChipsAsync()
-    {
-        await UsePrepareFoodOrderProcessSingleItemAsync(FoodItem.FishAndChips);
-    }
-
-    protected async Task UsePrepareFoodOrderProcessSingleItemAsync(FoodItem foodItem)
-    {
-        Kernel kernel = CreateKernelWithChatCompletion();
-        KernelProcess kernelProcess = SingleFoodItemProcess.CreateProcess().Build();
-
-        using var runningProcess = await kernelProcess.StartAsync(kernel, new KernelProcessEvent()
-        {
-            Id = SingleFoodItemProcess.ProcessEvents.SingleOrderReceived,
-            Data = foodItem
         });
     }
 }
